Add support for mouse scroll wheel scrolling
Each 'click' of the scroll wheel moves one line at a time, just as it
does in xterm. sam already has existing features to jump in larger
units (eg PgUp/PgDown), so I (cks) prefer this better. The
implementation is somewhat of a hack, partly because I'm not familiar
with samterm internals.