samterm: fix a panic in splitsect() due to buffered input
If you're at the end of a buffer and you force samterm to process
multiple typed characters at once where the last character is a
backspace, samterm will panic in splitsect(); see Github issue #88 for
a detailed discussion. The ultimate cause appears to be processing
special characters before we handle the input of regular
characters. To fix this, we move processing of Kcommand actions in
type() to after any accumulated regular characters have been
inserted. This mimics the approach type() used to process characters
in the pre-changes ('legacy') samterm.